+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Unreleased
 
 ### Changed
+- Neo block type settings are now rendered in PHP/Twig instead of JavaScript
 - Neo input blocks now show the block type's handle next to the name on desktops/tablets, for users who have their 'Show field handles in edit forms' user setting enabled
 
 ## 3.0.6 - 2022-06-02